+++ CHANGES.md
@@ -16,6 +16,13 @@ It is important to distinguish EAP from traditional pre-release software.
 Please note that the quality of EAP versions may at times be way below even
 usual beta standards.
 
+To Be Released
+--------------
+
+Bug fixes:
+
+* VIM-1098 Don't start visual selection when mouse click was actually drag over single character
+
 
 0.46, 2016-07-07
 ----------------